requirements.txt
# Core
streamlit>=1.35.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0

# Embeddings
sentence-transformers>=2.7.0

# Vector store
faiss-cpu>=1.8.0

# Document parsing
pdfplumber>=0.10.0
python-docx>=1.1.0

# Ollama Python client
ollama>=0.2.0
